Diseño de los botones y fetch listo
