+++ src/front/routes.jsx
@@ -8,7 +8,7 @@ import {
 import { Layout } from "./pages/Layout";
 import { Home } from "./pages/Home";
 import { Single } from "./pages/Single";
-import { Demo } from "./pages/Demo";
+import { Register } from "./pages/Register";
 // import { RegistroCarrier } from "./pages/RegistroCarrier";
 
 export const router = createBrowserRouter(
@@ -25,7 +25,7 @@ export const router = createBrowserRouter(
         {/* Nested Routes: Defines sub-routes within the BaseHome component. */}
         <Route path= "/" element={<Home />} />
         <Route path="/single/:theId" element={ <Single />} />  {/* Dynamic route for single items */}
-        <Route path="/demo" element={<Demo />} />
+        <Route path="/register" element={<Register />} />
         {/* <Route path="/RegistroCarrier" element={<RegistroCarrier />} /> */}
       </Route>
     )
